+++ src/backend/Storage/ExprCalc.Storage/Resources/SqliteQueries/SqliteDbQueryProvider.cs
@@ -69,6 +69,11 @@ namespace ExprCalc.Storage.Resources.SqliteQueries
                         CreatedAt DESC,
                         State
                     ) WHERE State != {(int)CalculationState.Success};
+
+                    CREATE INDEX IF NOT EXISTS idx_Calculations_CreatedAt_CalcResult ON Calculations (
+                        CreatedAt DESC,
+                        CalcResult
+                    ) WHERE CalcResult IS NOT NULL;
                     """;
 
                 command.ExecuteNonQuery();
@@ -261,6 +266,22 @@ namespace ExprCalc.Storage.Resources.SqliteQueries
                     conditionBuilder.Append(" AND ");
                 conditionBuilder.Append("Expression LIKE @Expression");
                 command.Parameters.Add("@Expression", SqliteType.Text).Value = "%" + filters.Expression + "%";
+                appendAnd = true;
+            }
+            if (filters.CalculationResultMin != null)
+            {
+                if (appendAnd)
+                    conditionBuilder.Append(" AND ");
+                conditionBuilder.Append("CalcResult IS NOT NULL AND CalcResult >= @CalculationResultMin");
+                command.Parameters.Add("@CalculationResultMin", SqliteType.Real).Value = filters.CalculationResultMin.Value;
+                appendAnd = true;
+            }
+            if (filters.CalculationResultMax != null)
+            {
+                if (appendAnd)
+                    conditionBuilder.Append(" AND ");
+                conditionBuilder.Append("CalcResult IS NOT NULL AND CalcResult < @CalculationResultMax");
+                command.Parameters.Add("@CalculationResultMax", SqliteType.Real).Value = filters.CalculationResultMax.Value;
                 appendAnd = true;
             }